se actualizo el archivo catwizzard_master.sql con la tabla recursiva, se creo modifico el objeto de dominio y se cambio el nombre de la  carpeta del proyecto de hello a io/catwizard
