+++ src/main/java/io/catwizard/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package io.catwizard;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
