+++ src/main/java/io/catwizard/DAO/CatwProductDao.java
@@ -1,6 +1,6 @@
-package hello.DAO;
+package io.catwizard.DAO;
 
-import hello.domain.Contact;
+import io.catwizard.domain.CatwProduct;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,5 +11,5 @@ import javax.transaction.Transactional;
  */
 @Transactional
 @Repository
-public interface ContactDao extends CrudRepository<Contact, Integer> {
+public interface CatwProductDao extends CrudRepository<CatwProduct, Integer> {
 }
