+++ src/main/java/io/catwizard/DAO/ContactDao.java
@@ -1,6 +1,6 @@
-package hello.DAO;
+package io.catwizard.DAO;
 
-import hello.domain.Server;
+import io.catwizard.domain.Contact;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,5 +11,5 @@ import javax.transaction.Transactional;
  */
 @Transactional
 @Repository
-public interface ServerDao extends CrudRepository<Server, Integer> {
+public interface ContactDao extends CrudRepository<Contact, Integer> {
 }
