+++ src/main/java/io/catwizard/DAO/ServerDao.java
@@ -1,6 +1,6 @@
-package hello.DAO;
+package io.catwizard.DAO;
 
-import hello.domain.Contract;
+import io.catwizard.domain.Server;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,5 +11,5 @@ import javax.transaction.Transactional;
  */
 @Transactional
 @Repository
-public interface ContractDao extends CrudRepository<Contract, Integer> {
+public interface ServerDao extends CrudRepository<Server, Integer> {
 }
