+++ src/main/java/io/catwizard/controller/GreetingController.java
@@ -1,8 +1,8 @@
-package hello.controller;
+package io.catwizard.controller;
 
 import java.util.concurrent.atomic.AtomicLong;
 
-import hello.domain.Greeting;
+import io.catwizard.domain.Greeting;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
